refactor(onboard): extract executeSandboxCommandForVerification to onboard/
